Move and rename show cmd runner
- Moved the runner out of the cmd/enaml package since it made the compilation command line more complicated and bloated the main package
